fix: enable direct download
