+++ packages/cli-plugin-scaffold-extensions/src/index.ts
@@ -1,37 +1,12 @@
-import { CliCommandScaffoldTemplate, PackageJson } from "@webiny/cli-plugin-scaffold/types";
-import os from "os";
+import { CliCommandScaffoldTemplate } from "@webiny/cli-plugin-scaffold/types";
 import path from "path";
-import util from "util";
-import ncpBase from "ncp";
-import readJson from "load-json-file";
-import writeJson from "write-json-file";
-import execa from "execa";
 import Case from "case";
-import { replaceInPath } from "replace-in-path";
-import WebinyError from "@webiny/error";
 import validateNpmPackageName from "validate-npm-package-name";
-import downloadFolderFromGitHub from "github-directory-downloader";
 import fs from "node:fs";
-import fsAsync from "node:fs/promises";
 
-/**
- * TODO: rewrite cli into typescript
- */
-// @ts-expect-error
-import { getProject, log } from "@webiny/cli/utils";
-import { generators } from "./generators";
-import {linkAllExtensions} from "~/generators/utils/linkAllExtensions";
-
-const ncp = util.promisify(ncpBase.ncp);
-
-interface Input {
-    type: string;
-    name: string;
-    packageName: string;
-    location: string;
-    dependencies?: string;
-    download?: string;
-}
+import { downloadAndLinkExtension } from "./downloadAndLinkExtension";
+import { generateExtension } from "./generateExtension";
+import { Input } from "./types";
 
 const EXTENSIONS_ROOT_FOLDER = "extensions";
 
@@ -134,153 +109,15 @@ export default () => [
                     }
                 ];
             },
-            generate: async ({ input, ora, context }) => {
-                const project = getProject();
-
-                try {
-                    let location;
-                    if (input.download) {
-                        let url = input.download;
-                        if (!url.startsWith("https")) {
-                            url = "https://github.com/webiny/webiny-examples/tree/master/" + url;
-                        }
-
-                        const randomId = String(Date.now());
-                        const tempDownloadFolderPath = path.join(
-                            os.tmpdir(),
-                            `wby-ext-${randomId}`
-                        );
-                        await downloadFolderFromGitHub(url, tempDownloadFolderPath, {
-                            muteLog: true,
-                            requests: 2
-                        });
-
-                        const extensionsFolderToCopy = path.join(
-                            tempDownloadFolderPath,
-                            "extensions"
-                        );
-                        await fsAsync.cp(extensionsFolderToCopy, EXTENSIONS_ROOT_FOLDER, {
-                            recursive: true
-                        });
-
-                        await linkAllExtensions();
-                    } else {
-                        const { type, name } = input;
-                        if (!type) {
-                            throw new Error("Missing extension type.");
-                        }
-
-                        const templatePath = path.join(__dirname, "templates", type);
-                        const templateExists = fs.existsSync(templatePath);
-                        if (!templateExists) {
-                            throw new Error("Unknown extension type.");
-                        }
-
-                        if (!name) {
-                            throw new Error("Missing extension name.");
-                        }
-
-                        let { packageName, location } = input;
-                        if (!packageName) {
-                            packageName = Case.kebab(name);
-                        }
-
-                        if (!location) {
-                            location = `${EXTENSIONS_ROOT_FOLDER}/${name}`;
-                        }
-
-                        if (fs.existsSync(location)) {
-                            throw new WebinyError(
-                                `The target location already exists "${location}"`
-                            );
-                        }
-
-                        ora.start(`Creating ${log.success.hl(name)} extension...`);
-
-                        // Copy template files
-                        fs.mkdirSync(location, { recursive: true });
-                        await ncp(templatePath, location);
-
-                        const baseTsConfigFullPath = path.resolve(project.root, "tsconfig.json");
-                        const baseTsConfigRelativePath = path.relative(
-                            location,
-                            baseTsConfigFullPath
-                        );
-
-                        const codeReplacements = [
-                            { find: "PACKAGE_NAME", replaceWith: packageName },
-                            {
-                                find: "BASE_TSCONFIG_PATH",
-                                replaceWith: baseTsConfigRelativePath
-                            }
-                        ];
-
-                        replaceInPath(path.join(location, "**/*.*"), codeReplacements);
-
-                        if (input.dependencies) {
-                            const packageJsonPath = path.join(location, "package.json");
-                            const packageJson = await readJson<PackageJson>(packageJsonPath);
-                            if (!packageJson.dependencies) {
-                                packageJson.dependencies = {};
-                            }
-
-                            const packages = input.dependencies.split(",");
-                            for (const packageName of packages) {
-                                const isWebinyPackage = packageName.startsWith("@webiny/");
-                                if (isWebinyPackage) {
-                                    packageJson.dependencies[packageName] = context.version;
-                                    continue;
-                                }
-
-                                try {
-                                    const { stdout } = await execa("npm", [
-                                        "view",
-                                        packageName,
-                                        "version",
-                                        "json"
-                                    ]);
-
-                                    packageJson.dependencies[packageName] = `^${stdout}`;
-                                } catch (e) {
-                                    throw new Error(
-                                        `Could not find ${log.error.hl(
-                                            packageName
-                                        )} NPM package. Please double-check the package name and try again.`,
-                                        { cause: e }
-                                    );
-                                }
-                            }
-
-                            await writeJson(packageJsonPath, packageJson);
-
-                            // Add package to workspaces
-                            const rootPackageJsonPath = path.join(project.root, "package.json");
-                            const rootPackageJson = await readJson<PackageJson>(rootPackageJsonPath);
-                            if (!rootPackageJson.workspaces.packages.includes(location)) {
-                                rootPackageJson.workspaces.packages.push(location);
-                                await writeJson(rootPackageJsonPath, rootPackageJson);
-                            }
-
-                            if (typeof generators[type] === "function") {
-                                await generators[type]({ input: { name, packageName } });
-                            }
-                        }
-                    }
-
-                    // Sleep for 1 second before proceeding with yarn installation.
-                    await new Promise(resolve => {
-                        setTimeout(resolve, 1000);
-                    });
-
-                    // Once everything is done, run `yarn` so the new packages are installed.
-                    await execa("yarn");
-
-                    ora.succeed(`New extension created in ${log.success.hl(location)}.`);
-                } catch (err) {
-                    ora.fail("Could not create extension. Please check the logs below.");
-                    console.log();
-                    console.log(err);
+            generate: async params => {
+                // The `templateArgs` is used by this scaffold to identify if the user wants
+                // to download an extension from the Webiny examples repository.
+                const downloadExtensionSource = params.input.templateArgs;
+                if (downloadExtensionSource) {
+                    return downloadAndLinkExtension(params);
                 }
+
+                return generateExtension(params);
             }
         }
     } as CliCommandScaffoldTemplate<Input>
