+++ packages/cli-plugin-scaffold-extensions/src/types.ts
@@ -6,3 +6,12 @@ interface PluginGeneratorParams {
 }
 
 export type PluginGenerator = (params: PluginGeneratorParams) => Promise<void>;
+
+export interface Input {
+    type: string;
+    name: string;
+    packageName: string;
+    location: string;
+    dependencies?: string;
+    templateArgs?: string;
+}
